fix(042): the correlator was SATURATING -- q = 1.00 was mostly an artifact, not a measurement
Both of the correlator's accumulators were int32 with saturating clamps, and at
the real operating point both of them railed. Measured on static_ir.bcnr:

  corr railed at INT32_MAX   73% of slot-ticks
  energy railed              80%
  BOTH railed together       73%
  q == 1.000                 75%  -- of which 98% were the both-railed case

When both rail, q = |corr|*256/energy = 2^31*256/2^31 = 256 exactly. So three
quarters of all q readings were the double-saturation artifact, and q is what
gates PROMOTION, HOLD ENTRY and the SCALE LADDER. It is also why q flipped
0.87 / 0.51 / 1.00 on consecutive ticks of a STATIC target.

The magnitudes are structural, not a corner case. extract() forms
(sum << 14) / (exposure_us * gain_q8 / 256); at 45 us / gain 1.0 that is a x364
multiplier, and the aperture SUM over ~1000 native px at ~2.4 frames/chip
reaches ~1e9 per bin before devs() multiplies by another 256. Measured true
maxima once widened: |corr| 2.15e9, energy 4.19e9 -- the latter nearly TWICE
what an int32 holds. dev[] was the same overflow one step earlier: computed in
int64, then truncated into an int32 slot.

Widened corr/energy/level/dev to int64 end to end, and the per-pixel position
surface with them -- truncating pr.corr into an int32 surf[] would have
reintroduced the identical bug one step later. Costs one extra scratch array
(124 * 8 B) and nothing measurable in time.

MEASURED after:  corr railed 0, q == 1.000 down to 22%, q p10/p50 0.922/0.996.

  static_ir      MEASURED fix   80% -> 98%   (LOCK 98% -> 99%)
  pend3          fixes 2593 -> 3479 (+34%),  MEASURED 54% -> 73%
                 0-1 76->97  1-2 72->93  2-3 63->86  3-5 52->81
                 5-8 33->55  >=8 16->21   relock p50 74 -> 42 ms
  pend_ir        fixes 2306 -> 2954 (+28%),  3-5 76->82  5-8 55->60

HONEST CAVEAT, not yet resolved: the false-lock share ROSE -- pend_ir 0.1% ->
2.2% (2 -> 65), pend3 0.2% -> 0.9%. More tracks survive and q is now a true
ratio rather than a railed constant, so sources that used to be indistinguishable
at q=1.00 now score their real value and some clear the promotion bar.

This is exactly the ground T085's min_mod_depth gate was meant to cover, and it
matters that ITS CHARACTERISATION WAS DONE ON SATURATED DATA: the "lamp p95 0.88
vs beacon median 1.9-22.6" separation was computed with energy railed, so those
numbers describe a clamp, not the scene. Re-measuring that separation on honest
energy is the next step and is now possible for the first time.

+++ firmware/beacon-receiver/src/app/beacon_trackd.cc
@@ -113,9 +113,10 @@ static void emit_diag_line(const Engine *eng)
         if (!s->used) continue;
         n += snprintf(buf + n, sizeof buf - (size_t)n,
                       "%s{\"slot\":%u,\"code\":%u,\"role\":%u,\"state\":%u,\"corr\":%d,"
-                      "\"energy\":%d,\"level\":%d,\"q\":%.3f,\"x\":%.2f,\"y\":%.2f,\"scale\":%u}",
+                      "\"energy\":%lld,\"level\":%lld,\"q\":%.3f,\"x\":%.2f,\"y\":%.2f,\"scale\":%u}",
                       first ? "" : ",", i, s->trk.code_id, s->role, s->trk.state,
-                      s->trk.last_corr, s->trk.last_energy, s->trk.last_level, s->trk.q_q8 / 256.0,
+                      (long long)s->trk.last_corr, (long long)s->trk.last_energy,
+                      (long long)s->trk.last_level, s->trk.q_q8 / 256.0,
                       s->trk.x_q8 / 256.0, s->trk.y_q8 / 256.0, s->trk.scale);
         first = 0;
     }
